helios_requirements.txt
huggingface_hub==0.27.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==1.26.4
pillow==10.4.0
# PyYaml==0.2.5
hydra-core==1.3.2
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.0.13
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
datasets==2.21.0
transformers==4.41.2
accelerate==0.34.2
mediapy
umap-learn
seaborn
opencv-python
wandb
-e diffusers
dreamsim
peft==0.11.0
huggingface_hub<1.0